// Client/src/interfaces/Media/TmdbTvDetails.ts
export interface TmdbTvDetails {
  adult?: boolean;
  backdrop_path?: string | null;
  created_by?: Array<{
    id: number;
    credit_id?: string;
    name: string;
    original_name?: string;
    gender?: number | null;
    profile_path?: string | null;
  }>;
  episode_run_time?: number[];
  first_air_date?: string | null;
  genres?: Array<{ id: number; name: string }>;
  homepage?: string | null;
  id: number;
  in_production?: boolean;
  languages?: string[];
  last_air_date?: string | null;
  last_episode_to_air?: {
    id?: number;
    name?: string;
    overview?: string;
    vote_average?: number;
    vote_count?: number;
    air_date?: string | null;
    episode_number?: number;
    episode_type?: string;
    production_code?: string;
    runtime?: number | null;
    season_number?: number;
    show_id?: number;
    still_path?: string | null;
  } | null;
  name?: string;
  next_episode_to_air?: {
    id?: number;
    name?: string;
    overview?: string;
    vote_average?: number;
    vote_count?: number;
    air_date?: string | null;
    episode_number?: number;
    episode_type?: string;
    production_code?: string;
    runtime?: number | null;
    season_number?: number;
    show_id?: number;
    still_path?: string | null;
  } | null;
  networks?: Array<{
    id?: number;
    logo_path?: string | null;
    name?: string;
    origin_country?: string;
  }>;
  number_of_episodes?: number;
  number_of_seasons?: number;
  origin_country?: string[];
  original_language?: string;
  original_name?: string;
  overview?: string;
  popularity?: number;
  poster_path?: string | null;
  production_companies?: Array<{
    id?: number;
    logo_path?: string | null;
    name?: string;
    origin_country?: string;
  }>;
  production_countries?: Array<{
    iso_3166_1?: string;
    name?: string;
  }>;
  seasons?: Array<{
    air_date?: string | null;
    episode_count?: number;
    id?: number;
    name?: string;
    overview?: string;
    poster_path?: string | null;
    season_number?: number;
    vote_average?: number;
  }>;
  spoken_languages?: Array<{
    english_name?: string;
    iso_639_1?: string;
    name?: string;
  }>;
  status?: string;
  tagline?: string | null;
  type?: string;
  vote_average?: number;
  vote_count?: number;
}